# internal/docs/openapi.yaml
openapi: 3.1.0
info:
  title: zoekt-server API
  version: 1.0.0
  description: |
    Code search server combining [Zoekt](https://github.com/sourcegraph/zoekt)
    with MCP (Model Context Protocol) for LLM tool use.

    Includes the upstream Zoekt search/list API, file viewer, reindexing, and MCP endpoints.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/file:
    get:
      tags: [Files]
      summary: Get file contents
      operationId: getFile
      description: |
        Retrieve file contents from a bare git repository at HEAD.
        Uses literal (not regex) repo and path parameters.
      parameters:
        - name: repo
          in: query
          required: true
          schema:
            type: string
          description: Full repository name
          example: "github.com/sourcegraph/zoekt"
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: File path within the repository
          example: "README.md"
      responses:
        "200":
          description: File contents as plain text
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Missing parameters
        "404":
          description: File or repository not found
  /api/reindex:
    post:
      tags: [Reindex]
      summary: Trigger reindex
      operationId: reindex
      description: |
        Queue a repository for immediate reindexing. The request is added to
        the high-priority queue and processed ahead of periodic indexing.
        The bare repo must already exist on disk (cloned by the mirror process).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repo]
              properties:
                repo:
                  type: string
                  description: "Repository path in hostname/org/name format"
                  example: "github.com/myorg/myrepo"
      responses:
        "202":
          description: Reindex task accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: Task ID (UUID)
                  status:
                    type: string
                    enum: [pending]
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Index queue full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...